Define CF_RETURN_* macros on CFBase
Clang allows for cf_returns_retained and cf_returns_not_retained
attributes on CoreFoundation-family functions to annotate whether they
return an object that must later be released or not. The macros
CF_RETURNS_RETAINED and CF_RETURNS_NOT_RETAINED are easier ways to
specify this attribute.

This commit defines these macros on CoreBase for increased
compatibility.