runtime: implement ARC bridging functions
Automatic Reference Counting (ARC) can only handle the memory management
of Objective-C objects and not that of C structs. To preserve the ease
of toll-free bridging, ARC has added the __bridge_retained and
__bridge_transfer directives for casting between these two types.

__bridge_retained specifies that an Objective-C object should be cast to
a C struct and that the struct needs to be manually released.
__bridge_transfer works the other way, and turns a C struct that would
need a manual release into an object that is now managed by ARC and no
longer needs that.

The Apple CoreFoundation API specifies the CFBridgingRetain() and
CFBridgingRelease() functions that do, respectively, the same things
these directives do. This commit implements these two functions exactly
as the compiler would handle the directives.

Co-authored-by: Stefan Bidigaray <stefanbidi@gmail.com>