GSCharacter: fix bug in conversion from UTF8 to UTF16
In the valid case that we passed a destination parameter to
GSUnicodeFromEncoding() that was smaller than the bytes read in the
source, CoreBase should continue reading the source and just not write
to the destination buffer.

GSUTF16CharacterAppend() did not guarantee that when converting from
UTF8 this would be the case, which caused a segfault due to a buffer
overflow.

This commit amends the issue.